Change a typo in the Dietary Restrictions package
