+++ src/main/java/app/AppBuilder.java
@@ -5,7 +5,7 @@ import java.awt.CardLayout;
 import javax.swing.JFrame;
 import javax.swing.JPanel;
 
-import data_access.dietery_restriction.DietResDataAccessObject;
+import data_access.dietary_restriction.DietResDataAccessObject;
 import data_access.favorites.FavoriteDataAccessObject;
 import data_access.generate_recipe.MealDbRecipeDetailsGateway;
 import data_access.generate_recipe.MealDbRecipeGateway;
@@ -14,22 +14,22 @@ import data_access.test_DAO.InMemoryInventoryReader;
 import data_access.user_recipe.FileDataAccessObject;
 import interface_adapter.DietResViewManagerModel;
 import interface_adapter.UserRecipesViewManagerModel;
-import interface_adapter.dietery_restriction.add_diet_res.AddDietResController;
-import interface_adapter.dietery_restriction.add_diet_res.AddDietResPresenter;
-import interface_adapter.dietery_restriction.add_diet_res.AddDietResViewModel;
+import interface_adapter.dietary_restriction.add_diet_res.AddDietResController;
+import interface_adapter.dietary_restriction.add_diet_res.AddDietResPresenter;
+import interface_adapter.dietary_restriction.add_diet_res.AddDietResViewModel;
 import interface_adapter.favorites.add_favorite.AddFavoriteController;
 import interface_adapter.favorites.add_favorite.AddFavoritePresenter;
 import interface_adapter.favorites.add_favorite.AddFavoriteViewModel;
 import interface_adapter.generate_recipe.generate_with_inventory.GenerateWithInventoryController;
 import interface_adapter.generate_recipe.generate_with_inventory.GenerateWithInventoryPresenter;
 import interface_adapter.generate_recipe.generate_with_inventory.GenerateWithInventoryViewModel;
-import interface_adapter.dietery_restriction.remove_diet_res.RemoveDietResController;
-import interface_adapter.dietery_restriction.remove_diet_res.RemoveDietResPresenter;
-import interface_adapter.dietery_restriction.remove_diet_res.RemoveDietResViewModel;
-import interface_adapter.dietery_restriction.view_diet_res.DietResViewModel;
-import interface_adapter.dietery_restriction.view_diet_res.DietResWindowModel;
-import interface_adapter.dietery_restriction.view_diet_res.ViewRestrictionsController;
-import interface_adapter.dietery_restriction.view_diet_res.ViewRestrictionsPresenter;
+import interface_adapter.dietary_restriction.remove_diet_res.RemoveDietResController;
+import interface_adapter.dietary_restriction.remove_diet_res.RemoveDietResPresenter;
+import interface_adapter.dietary_restriction.remove_diet_res.RemoveDietResViewModel;
+import interface_adapter.dietary_restriction.view_diet_res.DietResViewModel;
+import interface_adapter.dietary_restriction.view_diet_res.DietResWindowModel;
+import interface_adapter.dietary_restriction.view_diet_res.ViewRestrictionsController;
+import interface_adapter.dietary_restriction.view_diet_res.ViewRestrictionsPresenter;
 import interface_adapter.favorites.view_favorite.ViewFavoriteController;
 import interface_adapter.favorites.view_favorite.ViewFavoritePresenter;
 import interface_adapter.favorites.view_favorite.ViewFavoriteViewModel;
@@ -44,11 +44,11 @@ import interface_adapter.user_recipe.view_recipes.UserRecipesViewModel;
 import interface_adapter.user_recipe.view_recipes.ViewRecipesController;
 import interface_adapter.user_recipe.view_recipes.ViewRecipesPresenter;
 import use_case.user_recipe.add_recipe.add_ingredient.AddIngredientInteractor;
-import use_case.dietery_restriction.add_restrictions.AddDietResInteractor;
-import use_case.dietery_restriction.remove_restriction.RemoveDietResInteractor;
+import use_case.dietary_restriction.add_restrictions.AddDietResInteractor;
+import use_case.dietary_restriction.remove_restriction.RemoveDietResInteractor;
 import use_case.user_recipe.add_recipe.AddRecipeInteractor;
 import use_case.user_recipe.add_recipe.view_recipe_creator.ViewCreatorInteractor;
-import use_case.dietery_restriction.view_restrictions.ViewRestrictionsInteractor;
+import use_case.dietary_restriction.view_restrictions.ViewRestrictionsInteractor;
 import view.user_recipe_view.AddRecipeView;
 import use_case.favorites.add_favorite.AddFavoriteInteractor;
 import use_case.generate_recipe.generate_with_inventory.GenerateWithInventoryInputBoundary;
