+++ src/main/java/interface_adapter/dietary_restriction/view_diet_res/DietResWindowModel.java
@@ -1,0 +1,14 @@
+package interface_adapter.dietary_restriction.view_diet_res;
+
+import interface_adapter.ViewModel;
+
+public class DietResWindowModel extends ViewModel<Boolean> {
+
+    public static final String SET_VISIBLE = "set_visible";
+    public static final String VIEW_NAME = "dietary_restrictions_window";
+
+    public DietResWindowModel() {
+        super(VIEW_NAME);
+    }
+}
+
