--- src/main/java/interface_adapter/dietery_restriction/add_diet_res/AddDietResController.java
@@ -1,23 +1,0 @@
-package interface_adapter.dietery_restriction.add_diet_res;
-
-import entity.Ingredient;
-import use_case.dietery_restriction.add_restrictions.AddDietResInputBoundary;
-import use_case.dietery_restriction.add_restrictions.AddDietResInputData;
-
-public class AddDietResController {
-    private final AddDietResInputBoundary interactor;
-
-    public AddDietResController(AddDietResInputBoundary interactor){
-        this.interactor = interactor;
-    }
-
-    public void execute(Ingredient ingredient){
-        AddDietResInputData inputData = new AddDietResInputData(ingredient);
-        interactor.execute(inputData);
-    }
-
-    public void execute(String ingredientName){
-        AddDietResInputData inputData = new AddDietResInputData(ingredientName);
-        interactor.execute(inputData);
-    }
-}
