--- src/main/java/interface_adapter/dietery_restriction/add_diet_res/AddDietResViewModel.java
@@ -1,13 +1,0 @@
-package interface_adapter.dietery_restriction.add_diet_res;
-
-import interface_adapter.ViewModel;
-
-public class AddDietResViewModel extends ViewModel<AddDietResState> {
-    public static final String RESTRICTION_ADDED = "restriction_added";
-    public static final String RESTRICTION_ADD_FAILED = "restriction_add_failed";
-
-    public AddDietResViewModel() {
-        super("add_restriction");
-        setState(new AddDietResState());
-    }
-}
