--- src/main/java/interface_adapter/dietery_restriction/remove_diet_res/RemoveDietResController.java
@@ -1,23 +1,0 @@
-package interface_adapter.dietery_restriction.remove_diet_res;
-
-import entity.Ingredient;
-import use_case.dietery_restriction.remove_restriction.RemoveDietResInputBoundary;
-import use_case.dietery_restriction.remove_restriction.RemoveDietResInputData;
-
-public class RemoveDietResController {
-    private final RemoveDietResInputBoundary interactor;
-
-    public RemoveDietResController(RemoveDietResInputBoundary interactor) {
-        this.interactor = interactor;
-    }
-
-    /**
-     * Executes the remove restriction Use Case.
-     * @param ingredientName The name of the ingredient to remove.
-     */
-    public void execute(String ingredientName) {
-        Ingredient ingredient = new Ingredient(ingredientName, "");
-        RemoveDietResInputData inputData = new RemoveDietResInputData(ingredient);
-        interactor.execute(inputData);
-    }
-}
