--- src/main/java/interface_adapter/dietery_restriction/remove_diet_res/RemoveDietResState.java
@@ -1,24 +1,0 @@
-package interface_adapter.dietery_restriction.remove_diet_res;
-
-public class RemoveDietResState {
-    private String message = "";
-    private boolean isSuccess = false;
-
-    public RemoveDietResState() {}
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public boolean isSuccess() {
-        return isSuccess;
-    }
-
-    public void setSuccess(boolean success) {
-        isSuccess = success;
-    }
-}
