+++ src/main/java/use_case/dietary_restriction/view_restrictions/ViewRestrictionsInputBoundary.java
@@ -1,0 +1,9 @@
+package use_case.dietary_restriction.view_restrictions;
+
+public interface ViewRestrictionsInputBoundary {
+    /**
+     * Executes the "View Restrictions" Use Case.
+     * Triggers the retrieval of restricted ingredients and updates the view.
+     */
+    void execute();
+}
