+++ src/main/java/use_case/dietary_restriction/view_restrictions/ViewRestrictionsOutputBoundary.java
@@ -1,0 +1,9 @@
+package use_case.dietary_restriction.view_restrictions;
+
+public interface ViewRestrictionsOutputBoundary {
+    /**
+     * Prepares the success view with the list of restricted ingredients.
+     * @param outputData contains the list of restricted ingredients
+     */
+    void prepareSuccessView(ViewRestrictionsOutputData outputData);
+}
