--- src/main/java/use_case/dietery_restriction/add_restrictions/AddDietResInputBoundary.java
@@ -1,9 +1,0 @@
-package use_case.dietery_restriction.add_restrictions;
-
-public interface AddDietResInputBoundary {
-    /**
-     * Executes the add restriction use case.
-     * @param inputData the input data containing the ingredient to add to restrictions list
-     */
-    void execute(AddDietResInputData inputData);
-}
