--- src/main/java/use_case/dietery_restriction/remove_restriction/RemoveDietResInputBoundary.java
@@ -1,9 +1,0 @@
-package use_case.dietery_restriction.remove_restriction;
-
-public interface RemoveDietResInputBoundary {
-    /**
-     * Executes the remove restriction use case.
-     * @param inputData the input data containing the ingredient to remove from restrictions list
-     */
-    void execute(RemoveDietResInputData inputData);
-}
