--- src/main/java/use_case/dietery_restriction/remove_restriction/RemoveDietResInteractor.java
@@ -1,27 +1,0 @@
-package use_case.dietery_restriction.remove_restriction;
-
-import entity.Ingredient;
-import use_case.dietery_restriction.diet_res_ingredients.DietResDataAccessInterface;
-
-public class RemoveDietResInteractor implements RemoveDietResInputBoundary {
-    private final DietResDataAccessInterface dietResDataAccessObject;
-    private final RemoveDietResOutputBoundary removeDietResPresenter;
-
-    public RemoveDietResInteractor(DietResDataAccessInterface dietResDataAccessObject,
-                                   RemoveDietResOutputBoundary removeDietResPresenter) {
-        this.dietResDataAccessObject = dietResDataAccessObject;
-        this.removeDietResPresenter = removeDietResPresenter;
-    }
-
-    @Override
-    public void execute(RemoveDietResInputData inputData) {
-        Ingredient ingredient = inputData.getIngredient();
-
-        if (dietResDataAccessObject.isRestricted(ingredient)) {
-            dietResDataAccessObject.removeRestriction(ingredient);
-            removeDietResPresenter.prepareSuccessView();
-        } else {
-            removeDietResPresenter.prepareFailView("Ingredient not found in restrictions list.");
-        }
-    }
-}
