+++ src/main/java/window/DietResWindow.java
@@ -1,6 +1,6 @@
 package window;
 
-import interface_adapter.dietery_restriction.view_diet_res.DietResWindowModel;
+import interface_adapter.dietary_restriction.view_diet_res.DietResWindowModel;
 
 import javax.swing.*;
 import java.awt.*;
