+++ src/main/java/window/MainWindow.java
@@ -3,7 +3,7 @@ package window;
 import javax.swing.*;
 
 import interface_adapter.favorites.view_favorite.ViewFavoriteController;
-import interface_adapter.dietery_restriction.view_diet_res.ViewRestrictionsController;
+import interface_adapter.dietary_restriction.view_diet_res.ViewRestrictionsController;
 import interface_adapter.user_recipe.view_recipes.ViewRecipesController;
 import view.MainView;
